requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
idna==3.6
mccabe==0.7.0
pika==1.3.2
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
django-debug-toolbar==4.2.0
django-seed==0.3.1
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
